# SyntaxGraph: Usability and performance tweaks