Several small `SyntaxGraph` tweaks.  Some of these have been split out from #35 since they are tiny improvements that don't need to wait for data size experimentation.

Changes worth discussing:
## Do not coerce attributes to NamedTuple unnecessarily

It doesn't make sense to freeze Dict-attributes to NamedTuple after an `ensure_attributes` or `delete_attributes` call.  I assume this freezing was unintentional and not detected because the attribute storage type abstraction is implemented quite well.

Note that this change increases precompile time and decreases test time.  Below are outputs from `@time using JuliaLowering` and `@time include("test/runtests.jl")`

Before:
  - 45.395172 seconds (927.27 k allocations: 57.289 MiB, 0.08% gc time, 0.17% compilation time)
  - 24.653554 seconds (78.07 M allocations: 4.079 GiB, 4.25% gc time, 95.33% compilation time: <1% of which was recompilation)

After:
  - 59.053901 seconds (1.12 M allocations: 70.427 MiB, 0.06% gc time, 0.13% compilation time)
  - 11.100739 seconds (35.53 M allocations: 1.789 GiB, 4.09% gc time, 90.51% compilation time: 2% of which was recompilation)

(Our precompile statements could probably use some tweaking still)

## Delete ineffective freeze_attrs call

This post-parsing freeze call wasn't doing anything.  The freezing we got was actually from `ensure_attributes`.  The output of lowering now has Dict-attributes, which I think are fine, but if we want to freeze post-lowering we can do that.  Making the call effective has a bad effect on the test time above.

## Print more information when node does not have attribute (and no default is provided)

This is an unrecoverable error anyway, so print a bit more information about the node we tried to access.  Before:
  ```
  julia> st = jlower("function foo end")
  julia> st.value
  ERROR: Property `value[33]` not found
  ```
  After:
  ```
  julia> st.value
  ERROR: Property `value[33]` not found. Available attributes:
    kind = code_info,
    is_toplevel_thunk = true,
    source = 21,
    slots = JuliaLowering.Slot[]
  ```