@c42f 2025-08-28
> In my opinion, I think we could be even more aggressive:
> 
> ```julia
> mutable struct SyntaxGraph{Attrs <: Union{Dict{Symbol,Any},NamedTuple}}
> ```

It would be interesting if this helps, but surely it would be a sign that things aren't type inferred as expected and that should instead be fixed by a one or two well-placed type asserts?