@c42f 2025-08-29
> Sounds good; I'll merge this given the speedup and the easiness of switching the default back. I agree we do need to go back and look into performance further.

Sounds reasonable. Hopefully my attempts to make the `NamedTuple` version better will help ... we'll see.